fixes #8
Add a `--path` option in 'tasks create' to provide a path to a directory
which needs to be zipped instead of a zip file itself. The `--path` and `--code`
options are mutually exclusive.

`--path` will create a temporary file containing the zipped contents of
the directory at `--path` _including_ the directory itself (as this
seems to be the default zip behaviour) and upload this to the ExH
environment. Temp file is deleted afterwards.